+++ demos/assets.c
@@ -57,6 +57,7 @@
 #include "assets/inc/ui/scroll_view_h.data"
 #include "assets/inc/ui/scroll_bar_h.data"
 #include "assets/inc/ui/spinbox.data"
+#include "assets/inc/ui/label.data"
 #include "assets/inc/ui/gif_image.data"
 #include "assets/inc/ui/text_selector.data"
 #include "assets/inc/ui/big_font.data"
@@ -406,6 +407,7 @@ ret_t assets_init(void) {
   assets_manager_add(rm, ui_scroll_view_h);
   assets_manager_add(rm, ui_scroll_bar_h);
   assets_manager_add(rm, ui_spinbox);
+  assets_manager_add(rm, ui_label);
   assets_manager_add(rm, ui_gif_image);
   assets_manager_add(rm, ui_text_selector);
   assets_manager_add(rm, ui_big_font);